# swagger.yaml
openapi: 3.0.3
info:
  title: Projeto Locadora Roger Ribeiro - SENAI Jandira
  description: |-
    API destinada ao controle de filmes, atores, diretores, estúdios, personagens, faixas etárias e gêneros.
    Desenvolvida como projeto do SENAI Jandira - junto ao Professor Marcel.
  termsOfService: http://swagger.io/terms/
  contact:
    email: cjo.rogerribeiro@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
  - url: http://localhost:3030/v1/locadora
...
paths:
...
  /filme-ator/{id}:
    get:
      tags: [Filme Ator]
      summary: Busca relacionamento filme-ator por ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Relacionamento encontrado
    put:
      tags: [Filme Ator]
      summary: Atualiza relacionamento filme-ator
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilmeAtorInput'
      responses:
        '200':
          description: Atualizado com sucesso
    delete:
      tags: [Filme Ator]
      summary: Remove ator do filme
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Removido com sucesso
  /filme/{filmeId}/atores:
    get:
      tags: [Filme Ator]
      summary: Lista todos os atores de um filme
      parameters:
        - in: path
          name: filmeId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista de atores retornada
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AtorComPapel'
  /ator/{atorId}/filmes:
    get:
      tags: [Filme Ator]
      summary: Lista todos os filmes de um ator
      parameters:
        - in: path
          name: atorId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista de filmes retornada
  /filme-ator:
    post:
      tags: [Filme Ator]
      summary: Adiciona ator a um filme
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilmeAtorInput'
      responses:
        '201':
          description: Relacionamento criado com sucesso

  # Relacionamentos Filme-Diretores
...
  /filme-diretor/{id}:
    get:
      tags: [Filme Diretor]
      summary: Busca relacionamento filme-diretor por ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Relacionamento encontrado
    put:
      tags: [Filme Diretor]
      summary: Atualiza relacionamento filme-diretor
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilmeDiretorInput'
      responses:
        '200':
          description: Atualizado com sucesso
    delete:
      tags: [Filme Diretor]
      summary: Remove diretor do filme
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Removido com sucesso
  /filme/{filmeId}/diretores:
    get:
      tags: [Filme Diretor]
      summary: Lista todos os diretores de um filme
      parameters:
        - in: path
          name: filmeId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista de diretores retornada
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Diretor'
  /diretor/{diretorId}/filmes:
    get:
      tags: [Filme Diretor]
      summary: Lista todos os filmes de um diretor
      parameters:
        - in: path
          name: diretorId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista de filmes retornada
  /filme-diretor:
    post:
      tags: [Filme Diretor]
      summary: Adiciona diretor a um filme
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilmeDiretorInput'
      responses:
        '201':
          description: Relacionamento criado com sucesso

  # Relacionamentos Filme-Estúdios
...
  /filme-estudio/{id}:
    get:
      tags: [Filme Estúdio]
      summary: Busca relacionamento filme-estúdio por ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Relacionamento encontrado
    put:
      tags: [Filme Estúdio]
      summary: Atualiza relacionamento filme-estúdio
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilmeEstudioInput'
      responses:
        '200':
          description: Atualizado com sucesso
    delete:
      tags: [Filme Estúdio]
      summary: Remove estúdio do filme
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Removido com sucesso
  /filme/{filmeId}/estudios:
    get:
      tags: [Filme Estúdio]
      summary: Lista todos os estúdios de um filme
      parameters:
        - in: path
          name: filmeId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista de estúdios retornada
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EstudioComTipo'
  /estudio/{estudioId}/filmes:
    get:
      tags: [Filme Estúdio]
      summary: Lista todos os filmes de um estúdio
      parameters:
        - in: path
          name: estudioId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista de filmes retornada
  /filme-estudio:
    post:
      tags: [Filme Estúdio]
      summary: Adiciona estúdio a um filme
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilmeEstudioInput'
      responses:
        '201':
          description: Relacionamento criado com sucesso

  # Endpoints existentes
...
components:
  schemas:
...
    AtorComPapel:
      type: object
      properties:
        ator_id: { type: integer }
        nome: { type: string, maxLength: 150 }
        data_nascimento: { type: string, format: date }
        altura: { type: number, format: float }
        peso: { type: number, format: float }
        papel: { type: string, maxLength: 100 }
    Diretor:
      type: object
      required: [nome, data_nascimento]
      properties:
        diretor_id: { type: integer }
        nome: { type: string, maxLength: 120 }
        data_nascimento: { type: string, format: date }
        altura: { type: number, format: float }
        peso: { type: number, format: float }
...
    EstudioComTipo:
      type: object
      properties:
        estudio_id: { type: integer }
        nome: { type: string, maxLength: 120 }
        cnpj: { type: string, maxLength: 14 }
        logradouro: { type: string, maxLength: 120 }
        bairro: { type: string, maxLength: 50 }
        cidade: { type: string, maxLength: 20 }
        complemento: { type: string, maxLength: 20 }
        tipo_producao: { type: string, enum: [Principal, 'Co-produção', Distribuição] }
...
    FilmeAtorInput:
      type: object
      required: [filme_id, ator_id, papel]
      properties:
        filme_id: { type: integer }
        ator_id: { type: integer }
        papel: { type: string, maxLength: 100 }
    FilmeDiretorInput:
      type: object
      required: [filme_id, diretor_id]
      properties:
        filme_id: { type: integer }
        diretor_id: { type: integer }
    FilmeEstudioInput:
      type: object
      required: [filme_id, estudio_id, tipo_producao]
      properties:
        filme_id: { type: integer }
        estudio_id: { type: integer }
        tipo_producao: { type: string, enum: [Principal, 'Co-produção', Distribuição] }